♻️ refactor: redesinged the common modal for message and confirm
