+++ src/shared/components/modal/index.tsx
@@ -1,0 +1,2 @@
+export * from './Confirm';
+export * from './Message';
